refacto

+++ src/main/java/com/kobe/warehouse/service/facturation/dto/TiersPayantDossierFactureDto.java
@@ -1,5 +1,4 @@
 package com.kobe.warehouse.service.facturation.dto;
 
-import java.math.BigDecimal;
-
-public record TiersPayantDossierFactureDto(Long id, String name, BigDecimal totalAmount, int factureItemCount) {}
+public record TiersPayantDossierFactureDto(Long id, String name, Integer totalAmount, int factureItemCount) {
+}
